test: Create the relational database once for the whole unit suite (SDK-533)
api/test_live_events.py is the fourth directory that opens the SQLite
database without creating it -- and unlike the other three it fails even
alone on a clean checkout. One session-scoped fixture at cognee/tests/unit/
now creates the database and runs migrations before any unit test, replacing
the per-directory copies (the pre-existing one in modules/agents and the two
added earlier on this branch). Every shard is self-sufficient by
construction rather than by which directory happened to run first.

Verified: each of the four directories alone on a clean .cognee_system, and
the full suite through the same three least_duration shards CI runs, each
on a clean directory -- 1794 / 1813 / 1799 passed with no order-dependent
failures.

Refs SDK-533
